Refactor: Connect new api pricing endpoint as fallback
